Internal renaming and scope adjustments.
